+++ crates/ra_hir/src/lib.rs
@@ -24,7 +24,8 @@ pub mod source_binder;
 mod ids;
 mod macros;
 mod name;
-mod module;
+mod module_tree;
+mod nameres;
 mod function;
 mod adt;
 mod type_ref;
@@ -32,14 +33,13 @@ mod ty;
 mod impl_block;
 mod expr;
 
-pub mod code_model_api;
+mod code_model_api;
 mod code_model_impl;
 
 use crate::{
     db::HirDatabase,
     name::{AsName, KnownName},
     ids::{DefKind, SourceItemId, SourceFileItemId, SourceFileItems},
-    code_model_api::{Crate, CrateDependency}
 };
 
 pub use self::{
@@ -56,7 +56,10 @@ pub use self::{
 
 pub use self::function::FnSignatureInfo;
 
-pub use self::code_model_api::Module;
+pub use self::code_model_api::{
+    Crate, CrateDependency,
+    Module, ModuleSource, Problem,
+};
 
 pub enum Def {
     Module(Module),
